rabbit mq added

+++ slot-booking-service/src/main/java/com/stackroute/slotbookingservice/service/BookingServiceImpl.java
@@ -1,5 +1,7 @@
 package com.stackroute.slotbookingservice.service;
 
+import com.stackroute.slotbookingservice.exception.BookingAlreadyExist;
+import com.stackroute.slotbookingservice.exception.BookingNotFoundException;
 import com.stackroute.slotbookingservice.model.Booking;
 import com.stackroute.slotbookingservice.repository.BookingRepo;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -14,9 +16,13 @@ public class BookingServiceImpl implements BookingService {
     @Autowired
     private BookingRepo bookingRepo;
     @Override
-    public Booking saveData(Booking booking) {
-        return bookingRepo.save(booking);
-
+    public Booking saveData(Booking booking) throws BookingAlreadyExist {
+       if (bookingRepo.findById(booking.getAppointmentId()).isPresent()){
+           throw new BookingAlreadyExist("Booking is already exist");
+       }else
+        {
+            return bookingRepo.save(booking);
+        }
     }
 
     @Override
@@ -33,15 +39,38 @@ public class BookingServiceImpl implements BookingService {
     }
 
     @Override
-    public void deleteBooking(Integer appointmentId) {
-        bookingRepo.deleteById(appointmentId);
+    public void deleteBooking(Integer appointmentId) throws BookingNotFoundException {
+
+        if (bookingRepo.findById(appointmentId).isPresent()){
+            bookingRepo.deleteById(appointmentId);
+        }else
+        {
+             throw new BookingNotFoundException("Booking not exist");
+        }
+
 
     }
 
 
     @Override
-    public Booking updateBooking(Booking booking,Integer id) {
+    public Booking updateBooking(Booking booking,Integer id) throws BookingNotFoundException {
 
-        return bookingRepo.save(booking);
+        if (bookingRepo.findById(id).isPresent()) {
+            Booking booking1 = bookingRepo.findById(id).get();
+
+            booking1.setAddress(booking.getAddress());
+            booking1.setAppointmentDate(booking1.getAppointmentDate());
+            booking1.setCustomername(booking.getCustomername());
+            booking1.setMobile(booking.getMobile());
+            booking1.setAppointmentTime(booking.getAppointmentTime());
+            booking1.setEmailId(booking.getEmailId());
+            booking1.setCenterName(booking.getCenterName());
+
+          return  bookingRepo.save(booking1);
+        } else {
+            throw new BookingNotFoundException("booking not found ") ;
+        }
+
     }
+
 }
